new playability lists
- methodology is slightly different from previous attempts; now words
that have equal difference in equity get `(diff)/(number_of_words)`
bonus, not just (diff) as it was before. Also words played with the same
letters but lower in equity are discarded (e.g. if NIEAZOW(E) has 80
equity and WAZONIE has 70, only NIEAZOW(E) will be added to the
list). This will mostly affect multi-bingo racks
- `playability-pl-lemmas-sjp-new.txt` is a contracted list of base forms
from sjp.pl. Ideally we want base forms from OSPS, but I don't have
access to this
- `playability-pl-hooks.txt` is a list of hooks, but only those for
which the extended word and the base word have different
stems (e.g. (KOMPUTER)Y is not counted, but G(ŹLE) is)

The list is based on 94*3 runs of `test`, each with 50k games, which
gives 14m games in total. The seeds used were: 1-94, 1000-1094, 2000-2094.